refactor: rename previews
